chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.10",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
]
